requirements.txt
Django==1.9.8
django-allauth==0.28.0
django-bootstrap-form==3.2.1
django-bootstrap3==7.0.1
django-braces==1.10.0
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-filter==0.15.0
django-oauth-toolkit==0.10.0
django-rest-auth==0.8.1
django-serializers==0.6.0
djangorestframework==3.4.1
environ==1.0
httpie==0.9.4
oauthlib==1.0.3
Pillow==3.3.1
pi

python-openid==2.2.5
reportlab==3.3.0
requests-oauthlib==0.7.0
six==1.10.0